avg-visualization/requirements.txt
Django==3.1.14
django-cors-headers==3.11.0
djangorestframework==3.12.4
gunicorn==20.1.0
kafka-python==2.0.2
pandas